RESTAPI: add method connected
This method attempts to send a GET request for the url configured in the
PerfRepoRESTAPI object. If the requests succeeds then we assume that
we're connected otherwise not. This should include invalid urls.

In the future this should call a hello REST API method but that's not
supported by PerfRepo yet.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>