RESTAPI: improve url handling
Until now we had no checks of whether the specified url was correct.
This patch implements a check by using the urlparse python library. It
attempts to parse the url and if the scheme specified is not http or
https the url is rejected.

An additional change is that now the user can specify the url both with
and without the final '/' character as this was a common mistake that
I've observed. If it's not present I add it automatically. Because of
this I also had to change all the object url paths and method url paths
to be relative instead of absolute (leading with a '/').

Finally Rest method urls are now constructed using the urljoin method
from the urlparse library.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>