perfrepo-cli: use various error codes to distinguish errors
Currently any non-zero code returned from a command will
print the usage help message. This is not needed when the
error is caused by e.g. test was not found.

The patch adds several error codes to have better granularity
while reporting problem. The usage message will be printed out
only if the syntax is wrong.

Signed-off-by: Jan Tluka <jtluka@redhat.com>