add PerfRepoPermission class and related REST method
This class represents a permission setting attached to a PerfRepoReport
object. The setting defines a restriction to READ/WRITE permissions for
a specified GROUP/USER and can also be used to give PUBLIC access.

The related RESTAPI method allows us to add a single such object to a
PerfRepo instance.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>