PerfRepoTest: description is optional
Description of a Test is optional and doesn't hav eto be defined. If
such a test was returned by the REST API, xml parsing would fail with an
exception. This fixes the issue. Previously it was not possible to
create such a Test so this issue never showed up until now.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>