PerfRepoMetric: description is optional
Description of a Metric is optional and doesn't have to be defined. If
such a metric was returned by the REST API, xml parsing would fail with
an exception. This fixes the issue. Previously it was not possible to
create such a Metric so this issue never showed up until now.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>