remove StringType type checks
Same as with the NoneType, this is not the intended way to check for
string types.

For Python2 the check should be isinstance(x, basestring), which will
also include unicode strings. In our case we're more interested in
compatibility with Python3 so isinstance(x, str) is sufficient.

For Python3 we're interested in checking for the bytes type since that's
what gets returned by the requests library handling the http
communication for us.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>