python3 compatible imports
Some libraries were moved in Python3 which requires a change to the
import statements. Since it's just 2 libraries in our case we can keep
both imports with a try-except handling to support both Python2 and
Python3.

This should be the final change and the python-perfrepo library should
now be compatible with both versions of Python.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>